made AccountTable and AccountTableModel as Spring beans and change logic of initialize this classes in AccountPanel in AccountDao add new method getAllAccounts
