+++ src/main/java/com/account/AccountDaoImpl.java
@@ -1,10 +1,14 @@
 package com.account;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.stereotype.Repository;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.List;
 
 @Repository("accountDaoImpl")
 public class AccountDaoImpl implements AccountDao {
@@ -23,6 +27,20 @@ public class AccountDaoImpl implements AccountDao {
         String query = "delete from account where id = :p_account_id";
         SqlParameterSource namedParameter = new MapSqlParameterSource("p_account_id", accountId);
         namedParameterJdbcTemplate.update(query, namedParameter);
+    }
+
+    @Override
+    public List<Account> getAllAccounts() {
+        String query = "select * from account";
+        return namedParameterJdbcTemplate.query(query, new RowMapper<Account>() {
+            @Override
+            public Account mapRow(ResultSet rs, int rowNum) throws SQLException {
+                Account account = new Account();
+                account.setId(rs.getInt(1));
+                account.setAccountName(rs.getString(2));
+                return account;
+            }
+        });
     }
 
     @Autowired
