+++ src/main/java/com/account/panel/AccountPanel.java
@@ -3,17 +3,18 @@ package com.account.panel;
 import com.account.listener.AddAccountButtonListener;
 import com.account.listener.DeleteAccountButtonListener;
 import com.account.table.AccountTable;
+import com.account.table.AccountTableModel;
 import org.springframework.context.ApplicationContext;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import javax.swing.*;
 import java.awt.*;
 import static com.account.listener.AddAccountButtonListener.ADD_ACCOUNT_BUTTON_LISTENER_BEAN_NAME;
 import static com.account.listener.DeleteAccountButtonListener.DELETE_ACCOUNT_BUTTON_LISTENER_BEAN;
+import static com.account.table.AccountTable.ACCOUNT_TABLE_BEAN_NAME;
+import static com.account.table.AccountTableModel.ACCOUNT_TABLE_MODEL_BEAN_NAME;
 
 public class AccountPanel extends JPanel {
 
     private FlowLayout accountLayout = new FlowLayout();
-    private BorderLayout layout = new BorderLayout();
     private JLabel accountLabel = new JLabel("Enter account name");
     private JTextField accountInput = new JTextField("", 5);
     private JButton addAccountButton = new JButton("Add account");
@@ -21,6 +22,7 @@ public class AccountPanel extends JPanel {
     private AddAccountButtonListener addAccountButtonListener;
     private DeleteAccountButtonListener deleteAccountButtonListener;
     private AccountTable accountTable;
+    private AccountTableModel accountTableModel;
 
     public AccountPanel(ApplicationContext applicationContext) {
         addComponentsOnPanel(applicationContext);
@@ -32,14 +34,15 @@ public class AccountPanel extends JPanel {
     }
 
     public void addComponentsOnPanel(ApplicationContext applicationContext) {
-        NamedParameterJdbcTemplate namedParameterJdbcTemplate =
-                (NamedParameterJdbcTemplate) applicationContext.getBean("namedParameterJdbcTemplate");
+        accountTable = (AccountTable) applicationContext.getBean(ACCOUNT_TABLE_BEAN_NAME);
+        accountTableModel = (AccountTableModel) applicationContext.getBean(ACCOUNT_TABLE_MODEL_BEAN_NAME);
+        accountTable.setModel(accountTableModel);
+        accountTable.refreshModel();
         setLayout(accountLayout);
         add(accountLabel);
         add(accountInput);
         add(addAccountButton);
         add(deleteAccountButton);
-        accountTable = new AccountTable(namedParameterJdbcTemplate);
         JScrollPane jScrollPane = new JScrollPane();
         jScrollPane.setViewportView(accountTable);
         add(jScrollPane);
